Fix method invocation to work regardless of indirection.
The previous version has a bug wherein a method which takes a
pointer receiver will erroneously be called when the object checked
is a non-pointer. This is because `reflect` is aware of the method
receiver type when enumerating methods, and will not return any methods
of a different indirection level.

Since the old implementation only dereferenced values, any methods above
the current indirection level of the value queried would not be seen.

This patch adds new test cases which illustrate the bug, as well as new
function invocation code which checks for functions at a higher level of
indirection. It also removes a large chunk of (what I presume was) old
code to handle method invocations, but that was no longer used.